requirements.txt
django==3.0.6
django-ipware==3.0.2
xlwt==1.3.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
uwsgi
celery
redis
django_celery_beat